Add deep and light sleep modes
The setup function does everything we need it to due to deep sleep
function essentially resetting the device.
The code simply waits for the user to arm the device with a button,
enters light sleep and waits for someone to trip the alarm,
then it connects to WiFi and enters deep sleep.

We will now need to send an email with SMTP.
Then we need to add some robustness to check if the door is left open.